Масштабный рефакторинг
