+++ CodeColoring/CodeColoring/ArgsDecoder/DecodedArguments.cs
@@ -1,0 +1,15 @@
+﻿using CodeColoring.Colorizer;
+using CodeColoring.OutputFormat;
+using CodeColoring.ProgrammingLanguage;
+
+namespace CodeColoring.ArgsDecoder
+{
+    public class DecodedArguments
+    {
+        public string InputFilePath;
+        public ColorPalette ColorPalette;
+        public IProgrammingLanguage ProgrammingLanguage;
+        public string OutputFilePath;
+        public IOutputFormat OutputFormat;
+    }
+}
