+++ CodeColoring/CodeColoring/ArgsDecoder/IArgsDecoder.cs
@@ -1,21 +1,10 @@
-﻿using CodeColoring.Colorizer;
-using CodeColoring.OutputFormat;
-using CodeColoring.ProgrammingLanguage;
+﻿using System.Collections.Generic;
 
 namespace CodeColoring.ArgsDecoder
 {
     public interface IArgsDecoder
     {
         string Help { get; }
-        DecodedArguments Decode(string[] args);
-    }
-
-    public class DecodedArguments
-    {
-        public string InputFilePath;
-        public ColorPalette ColorPalette;
-        public IProgrammingLanguage ProgrammingLanguage;
-        public string OutputFilePath;
-        public IOutputFormat OutputFormat;
+        DecodedArguments Decode(IEnumerable<string> args);
     }
 }
