+++ CodeColoring/CodeColoring/Colorizer/ColorPalette.cs
@@ -16,7 +16,7 @@ namespace CodeColoring.Colorizer
         public Color BackgroundColor { get; protected set; }
         public string Name { get; protected set; }
 
-        public ColorPalette()
+        protected ColorPalette()
         {
             SymbolColor = Color.Black;
             FunctionDefinitionColor = Color.Black;
