+++ CodeColoring/CodeColoring/Colorizer/Colorizer.cs
@@ -6,7 +6,7 @@ namespace CodeColoring.Colorizer
 {
     public class Colorizer
     {
-        public Color Colorize(LanguageUnit unit, ColorPalette palette) =>
+        public static Color Colorize(LanguageUnit unit, ColorPalette palette) =>
             unit switch
             {
                 LanguageUnit.Variable => palette.VariableColor,
