+++ CodeColoring/CodeColoring/ProgrammingLanguage/LanguageUnit.cs
@@ -1,0 +1,15 @@
+﻿namespace CodeColoring.ProgrammingLanguage
+{
+    public enum LanguageUnit
+    {
+        Unknown,
+        Variable,
+        Symbol,
+        FunctionDefinition,
+        Function,
+        Operator,
+        Value,
+        Whitespace,
+        Comment
+    }
+}
