+++ CodeColoring/CodeColoring/ProgrammingLanguage/ParseUnit.cs
@@ -1,0 +1,14 @@
+﻿namespace CodeColoring.ProgrammingLanguage
+{
+    public readonly struct ParseUnit
+    {
+        public readonly string Symbol;
+        public readonly LanguageUnit Unit;
+
+        public ParseUnit(LanguageUnit unit, string symbol)
+        {
+            Unit = unit;
+            Symbol = symbol;
+        }
+    }
+}
