+++ CodeColoring/CodeColoring/ProgrammingLanguage/ParsingResult.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace CodeColoring.ProgrammingLanguage
+{
+    public class ParsingResult
+    {
+        public List<ParseUnit> Result = new();
+    }
+}
